+++ cardinal-components-entity/src/main/java/org/ladysnake/cca/mixin/entity/common/MixinEntity.java
@@ -54,7 +54,7 @@ public abstract class MixinEntity implements ComponentProvider {
     private ComponentContainer components;
 
     @Shadow
-    public World world;
+    private World world;
 
     @Shadow public abstract int getId();